### 2017-10-05 b1e6ba7
Fix #28 docker-compose worker
