### 2017-10-06 58b62ff
adding DEFAULT_PRIVATE setting
