+++ shub/apps/main/models.py
@@ -29,8 +29,11 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 '''
 
-from shub.settings import MEDIA_ROOT
-
+from shub.settings import (
+    MEDIA_ROOT,
+    PRIVATE_ONLY, 
+    DEFAULT_PRIVATE
+)
 from shub.apps.api.models import ImageFile
 from shub.apps.users.models import User
 from django.core.urlresolvers import reverse
@@ -70,8 +73,9 @@ ACTIVE_CHOICES = ((True, 'Active'),
 
 
 def get_privacy_default():
-    from shub.settings import PRIVATE_ONLY
-    return PRIVATE_ONLY
+    if PRIVATE_ONLY is True:
+        return PRIVATE_ONLY
+    return DEFAULT_PRIVATE
 
 
 def has_edit_permission(instance,request):
