### 2017-10-06 #419 @pjotrp
Dear @smoe would you care to review this Singularity submission?
