### 2017-10-06 #419 @smoe
Dear @pjotrp,

Singularity is important and everything that fosters its acceptance is appreciated.

The setup is technically sound. I like the distributed effort. What seems missing is the concept of mirroring images for long time availability when a site goes down, which then truly preserves setups for scientific reproducibility. For that to happen, we would need identifiers that are not URL-dependent, though. Another feature I am missing is the automated creation of images, which would nicely work together with the description of CWL workflows that features references to software registries which in turn started to have references to Linux distributions (OMICtools now, bio.tools soon, I hope) and Debian at least having references to the registries already.

Not having an exact clue what you ask me to review now, I went through your JOSS author guidelines on http://joss.theoj.org/about#author_guidelines:
 * License: BSD 3 clause - ok
 * paper.md is short, a bit too short IMHO, but present. Missing for my heart is a clarification how functionality and contributors differ from http://singularity.lbl.gov/, singularity-hub.org and github. Sorry, this is mostly trivial, but I strongly suggest to make these few lines as simple as possible. Frankly, I do not even expect everyone to know what a container is. Should users of docker also consider singularityhub instead of hub.docker.io? Are there features of singularity that are particularly supported by singularityhub?
 * authors found
 * The "summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience" as I already outlined above is marginal.
 * The "clear statement of need that illustrates the purpose of the software" we have - not that any fan of singularity would need it.
 * The "list of key references including a link to the software archive" is weak. Once the paper gets extended with the prior requested bit of extra content for those who are not using singularity already, this will become better. Maybe the authors' guide is a bit vague here. There is the citation system http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html that should be used.
 * Does it "Mention (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it"? No, it doesn't. I described in my initial lines who is waiting for it, though that is possibly too immature to be put into this paper.md. But you can of course outline that you foresee these kind of interactions between your service and others.

The other "requirements for submission" I find kind of redundant and I consider them to be fulfilled.

What I would typically do now is to install this locally and see if this works, which I kind of hope not to be imposed on me because the authors declare to us that they are working with exactly that setup themselves and that what we see is live on https://singularity-hub.org/collections or elsewhere - this is what should be explicitly stated in the paper.md, too.

Otherwise, I hope this was not too obvious, I am a fan.
