### 2017-10-06 #426 @whedon
[REVIEW]: Singularity Registry: Open Source Registry for Singularity Images
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://www.github.com/singularityhub/sregistry" target ="_blank">https://www.github.com/singularityhub/sregistry</a>
**Version:** v1.0.0
