### 2017-10-06 #426 @pjotrp
@smoe please tick the boxes, thanks! @vsoch please follow up.
