### 2017-10-06 #426 @vsoch
Thank you @smoe and @pjotrp for starting the discussion! This is my first review with JOSS so please tell me if I am doing something incorrectly. I will follow up here with answers to your questions, and then make a suggested list of changes for the repository. @whedon is asking for the issues to be in the target repository or linked, and so I will address the decided changes there.

>> Singularity is important and everything that fosters its acceptance is appreciated.

Thank you! I feel the same way.

## Mirroring and Backup

>> The setup is technically sound. I like the distributed effort. What seems missing is the concept of mirroring images for long time availability when a site goes down, which then truly preserves setups for scientific reproducibility.

I have the same sentiment that there is risk in any kind of infrastructure that is not properly mirrored. The challenge into integrating a mirror strategy into Singularity Registry is that it would be very challenging to impose a "one storage fits all" strategy. For example, by way of Docker the entire software can be brought up with:
```
docker-compose up -d
```

and this includes a postgres database image. However, the suggested deployment is to use a (non container) postgres image with something like a hot standby for backup, and then a mirror to some external storage (e.g., local filesystem) or cloud (e.g., object storage). What I think is in the scope of Singularity Registry is to provide the core software for organization, but leave the details about how backup is done to the user. This fits well with the two use cases that I see:
 - the single user: using the registry on a local machine, an image-based postgres is sufficient
 - an institution: an institution deployed registry would have the knowledge and background to create some database with backup.

In other words, the responsibility of ensuring the replication of the registry is going to differ based on the deployment, and is understood to be the choice of the deploying institution. What I've done to address back up, from the base that I can control (singularity-hub) is to have a deployment with a mirrored database, and all images in backed up cloud storage. If the web server goes down and the postgres and/or its backup endure, then it's not an issue to bring up the application again. If the web server and both databases go down, the entire set of containers and metadata in Storage can be used to repopulate a new database, and users are linked via the Github repos.

## Identifiers

>> For that to happen, we would need identifiers that are not URL-dependent, though.

Yes this is definitely the case! When a user references a Singularity Hub / Registry image, it might look something like this:

```
singularity pull shub://vsoch/hello-world
```

implicitly said in the above is the default url for the Singularity Hub registry - this produces the same outcome:

```
singularity pull shub://singularity-hub.org/vsoch/hello-world
```
and so if another registry was deployed with the same image uri, the registry name would need to be specified for the image.

```
singularity pull shub:// <registry> /vsoch/hello-world
```

@smoe do you have suggestions for how this could be better? For Singularity Hub, the namespace is well maintained to correspond with Github. For a Singularity Registry, the naming schema is up to the owner. For example, if I were running a cluster, I might have namespace correspond to user names and/or groups, and then projects owned by them. Each image does carry it's unique version (a hash of the file) however it wouldn't be intuitive to search and find. What (I think) you are hinting at is some sort of global wrapper to registries (akin to a package manager) that can find an image based on criteria, no matter where it is! This is a great idea, although would be a separate project (that uses the registries).

## Building
>> Another feature I am missing is the automated creation of images, which would nicely work together with the description of CWL workflows that features references to software registries which in turn started to have references to Linux distributions (OMICtools now, bio.tools soon, I hope) and Debian at least having references to the registries already.

Building is really important! And actually, I bit off more than I could chew a few years ago when my original vision for Singularity Hub was centered around the workflows. I realized we needed just storage and management before that.

To best answer this question, I want to make a distinction between Singularity Registry, and it's original base (the current Singularity Hub) that does in fact provide a building service. Singularity Hub was originally created to build images for users, as a free service, directly from a Github repository, and then have the images available immediately via the Singularity command line tool (for example, [here is the repository](github.com/vsoch/hello-world) that builds the image example above, when I push the build is done in the cloud automatically) and here is a quick image to show how it works:

![figure1-shub-flow](https://user-images.githubusercontent.com/814322/31297775-35a9173e-aa9c-11e7-9038-0aa780b21713.png)

 - I connect a repository to Singularity Hub
 - I push to the repository
 - The image is built and available via the singularity software

It's thus very possible for some higher level workflow tool to use Singularity Hub (or Singularity Registry) API to get the command to download an image. What is out of scope is actually running the workflow, and how it plugs into a workflow tool. [Here is an example](https://www.nextflow.io/blog/2016/more-fun-containers-hpc.html) with nextflow.

The Registry is a little bit of a different use case. I created the Registry because institutions / users came to me and said "I don't want to use Github, and I don't want even to share. I just want a manager for my institution images that I can push after build, and then pull from and be able to use on my resource." So Singularity Registry does not have builders or any cloud integrations - it's storage is a local file system as opposed to a cloud resource.

However, what I'm hoping to do for Singularity Hub and a new addition called [SCI-F apps](http://containers-ftw.org/apps/) (and this is probably outside of the scope of this review) is to have easy templates / buttons so that a user can easily plug it into a workflow.  [There are many workflows](https://github.com/pditommaso/awesome-pipeline) so likely I would start with major cloud providers. For apps, the idea is to provide easy templates ([for example, something to produce a SLURM job file](http://containers-ftw.org/apps/scif/hpc/slurm/hpc-slurm-submit/)) that helps with container integration.

## Singularity Hub vs. Registry vs. Github
- **Singularity Hub** is a registry that also provides a cloud build service for users, and it takes advantage of Github for ease of use. The user pushes to Github, a builder is deployed, and the image available to the user. Singularity Hub would allow a user to build and run an image from a resource where he or she doesn't have sudo simply by using Github as a middleman.
- **Singularity Registry** is similarly an image registry that plugs in natively to the singularity software, but it places no dependencies on Github, and puts the power of deciding how to build in the hands of the user. this could mean building after tests pass with a "push" command in a Github repo, building via a SLURM job, or on a private server. While Singularity Hub is entirely public and only allows for a minimum
number of private images, a Singularity Registry can be entirely private, with expiring tokens that can be shared.
- **Github**: Github serves as the easy connection to ensure version control of images. Each built container is linked to an exact commit, corresponding to the recipe and build dependencies for it. Singularity Registry does not enforce using Github, but encourages it. In discussion with a few centers, Github / GitLab (or other version control) seems to be common.

## Docker Hub
>> Should users of docker also consider singularityhub instead of hub.docker.io?
The two registries are similar in that they produce a container, however Docker Hub is fundamentally different in that it serves layers (.tar.gz) and not entire images. If one layer changes or is lost, the same image cannot be built. Singularity Registry and Hub serve entire images, no assembly required. I don't see the two as competitors, because many choose to use Docker Hub as a base to build a new image. I would say the decision of which to use is up to the user. I commonly build Docker images and then just pull them as singularity:

```
singularity run docker://godlovedc/lolcow
```
Any Docker image can be immediately converted to Singularity by adding the .tar.gz layers to an empty image base.  As far as I know, Docker doesn't have a nice conversion to get a Docker image from a Singularity one. Some users (myself included) find the Singularity recipe format to be more intuitive than Docker, and make the decision based on that.

## Are there features of singularity that are particularly supported by singularityhub?
Singularity aims to support scientific containers, so Singularity Hub and Registry take an extra step to serve metadata about the containers via the API. It's important to know about usage (downloads and stars) but also software, environment variables, labels, and runscripts. This supports being able to do more research analytics across containers to better understand how containers (and more broadly software) help answer scientific questions. Given the issues we have with reproducibility, I think this is essential

## Citations
Thank you for linking to the citation page. I'd be happy to add more references. Given that the description talks on a high level about Docker, containers, would the most appropriate references be to related web content? It's challenging to find papers that are specifically about container software and infrastructure - there are mostly analyses that use them.

## Ongoing Research Projects
I can definitely say that community members are using it, reflected in discussion on the list and Github issues, but I can't comment on specific projects.

# Submission Changes
I'll go through your comments and create a list of suggested changes. Please comment (add / change) any of those that you see fit:

- [ ] paper.md: add clarification for how functionality and contributors differ from http://singularity.lbl.gov/, singularity-hub.org and github
- [ ] paper.md: simplify the introduction and terms, and make a better distinction between Singularity and Docker. Better formulate "the high-level functionality and purpose of the software for a diverse, non-specialist audience" and a "clear statement of need that illustrates the purpose of the software"
- [ ] improve references with citation system: http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html and (additional references from web?)

Thank you for your thorough review! I can address the above points after you've taken a look, and I'll try to organize the issues and then fixes in context of the check boxes.
