### 2017-10-07 #426 @pjotrp
BTW have you thought of using IPFS as a backend? That could solve the redundancy question. We are experimenting with IPFS in several projects.
