### 2017-10-08 d23f418
updating paper, adding JOSS status badge
