+++ paper/paper.bib
@@ -54,3 +54,154 @@
   howpublished = "\url{https://github.com/singularityhub/sregistry}",
   note         = "Accessed: 2017-9-26"
 }
+
+
+
+@BOOK{Smith2005-kg,
+  title     = "Virtual Machines: Versatile Platforms for Systems and Processes",
+  author    = "Smith, J E and Nair, R",
+  publisher = "Morgan Kaufmann Publishers",
+  series    = "The Morgan Kaufmann Series in Computer Architecture and Design
+               Series",
+  year      =  2005
+}
+
+
+@ARTICLE{Merkel2014-da,
+  title     = "Docker: Lightweight Linux Containers for Consistent Development
+               and Deployment",
+  author    = "Merkel, Dirk",
+  journal   = "Linux J.",
+  publisher = "Belltown Media",
+  volume    =  2014,
+  number    =  239,
+  month     =  mar,
+  year      =  2014,
+  address   = "Houston, TX"
+}
+
+@MISC{noauthor_2015-ig,
+  title        = "Docker-based solutions to reproducibility in science - Seven
+                  Bridges",
+  booktitle    = "Seven Bridges",
+  abstract     = "Seven Bridges is launching a toolkit for creating fully
+                  portable bioinformatics workflows, using Docker and the
+                  Common Workflow Language.",
+  month        =  jun,
+  year         =  2015,
+  howpublished = "\url{https://blog.sbgenomics.com/docker-based-solutions-to-reproducibility-in-science/}",
+  note         = "Accessed: 2016-12-17"
+}
+
+@ARTICLE{Ali2016-rh,
+  title     = "The Case for Docker in Multicloud Enabled Bioinformatics
+               Applications",
+  author    = "Ali, Ahmed Abdullah and El-Kalioby, Mohamed and Abouelhoda,
+               Mohamed",
+  abstract  = "The introduction of next generation sequencing technologies did
+               not bring only huge amounts of biological data but also highly
+               sophisticated and versatile analysis workflows and systems.
+               These new cha",
+  journal   = "Conference on Bioinformatics and …",
+  publisher = "Springer International Publishing",
+  pages     = "587--601",
+  month     =  apr,
+  year      =  2016,
+  language  = "en"
+}
+
+@ARTICLE{Moreews2015-dy,
+  title       = "{BioShaDock}: a community driven bioinformatics shared
+                 Docker-based tools registry",
+  author      = "Moreews, Fran{\c c}ois and Sallou, Olivier and M{\'e}nager,
+                 Herv{\'e} and Le Bras, Yvan and Monjeaud, Cyril and Blanchet,
+                 Christophe and Collin, Olivier",
+  affiliation = "Genscale team, IRISA, Rennes, France. Genouest Bioinformatics
+                 Facility, University of Rennes 1/IRISA, Rennes, France. Centre
+                 d'Informatique pour la Biologie, C3BI, Institut Pasteur,
+                 Paris, France. Genouest Bioinformatics Facility, University of
+                 Rennes 1/IRISA, Rennes, France. Genouest Bioinformatics
+                 Facility, University of Rennes 1/IRISA, Rennes, France.
+                 Genouest Bioinformatics Facility, University of Rennes
+                 1/IRISA, Rennes, France. French Institute of Bioinformatics,
+                 CNRS IFB-Core, Gif-sur-Yvette, France.",
+  abstract    = "Linux container technologies, as represented by Docker,
+                 provide an alternative to complex and time-consuming
+                 installation processes needed for scientiﬁc software. The ease
+                 of deployment and the process isolation they enable, as well
+                 as the reproducibility they permit across environments and
+                 versions, are among the qualities that make them interesting
+                 candidates for the construction of bioinformatic
[…]
