+++ paper/paper.md
@@ -18,16 +18,20 @@ bibliography: paper.bib
 
 
 # Summary
+
+## A Need for Reproducible Science
 Using computational methods to answer scientific questions of interest is an important task to increase our knowledge about the world. Along with careful assembly of protocol and relevant datasets, the scientist that faces this challenging task must also write software to perform the analysis. Running the software in combination with the dataset is a typical recipe for scientific discovery. When such a result comes to fruition, the scientist writes it up for publication in a scientific journal. 
 
-The ability to replicate a result, then, can be thought of as increasing our confidence in the finding. The extent to which a published finding affords a second scientist to repeat the steps to achieve the result is called reproducibility. Reproducibility, in that it allows for repeated testing of an interesting question to validate knowledge about the world, is a foundation of science. While the original research can be an arduous task, often the culmination of years of work and committment, attemps to reproduce a series of methods to assess if the finding replicates is equally challenging. The researcher must minimally have enough documentation to describe the original data and methods to put together an equivalent experiment, and then uses computational methods to look for evidence to assert or reject the original hypothesis. 
+The ability to replicate a result, then, can be thought of as increasing our confidence in the finding. The extent to which a published finding affords a second scientist to repeat the steps to achieve the result is called reproducibility. Reproducibility, in that it allows for repeated testing of an interesting question to validate knowledge about the world, is a foundation of science. While the original research can be an arduous task, often the culmination of years of work and committment, attemps to reproduce a series of methods to assess if the finding replicates is equally challenging. The researcher must minimally have enough documentation to describe the original data and methods to put together an equivalent experiment, and then use computational methods to look for evidence to assert or reject the original hypothesis. 
 
 Unfortunately, many scientists are not able to provide the minimum product to allow others to reproduce their work. It could be an issue of time - the modern scientist is burdened with writing grants, managing staff, and fighting for tenure. It could be an issue of education. Graduate school training is heavily focused on a particular domain of interest, and developing skills to learn to program, use version control, and test is outside the scope of the program. It might also be entirely infeasible. If the experiments were run on a particular supercomputer and/or with a custom software stack, it is a non trivial task to provide that environment to others. The inability to easily share environments and software serves as a direct threat to scientific reproducibility.
 
+## Encapsulation of Environments with Containers
 The idea that the entire software stack, including libraries and specific versions of dependencies, could be put into a container and shared offered promise to help this problem. Linux containers, which can be thought of as encapsulated environments that host their own operating systems, software, and flie contents, were a deal breaker when coming onto the scene in early 2015 [@ARTICLE{Merkel2014-da]. Like Virtual Machines [@BOOK{Smith2005-kg], containers can make it easy to run a newer software stack on an older host, or to package up all the necessary software to run a scientific experiment, and have confidence that when sharing the container, it will run without a hitch. In early 2015, an early player on the scene, an enterprise container solution called Docker, started to be embraced by the scientific community [@MISC{noauthor_2015-ig;@ARTICLE{Ali2016-rh;@ARTICLE{Belmann2015-eb;@ARTICLE{Moreews2015-dy]. Docker containers were ideal for enterprise deployments, but posed huge security hazards if installed on a shared resource.
 
 It wasn't until the introduction of the Singularity software [@Kurtzer2017-x] that these workflows could be securely deployed on local cluster resources. For the first time, scientists could package up all of the software and libraries needed for their research, and deliver a complete package for a second scientist to reproduce the work. Singularity took the high performance computing world by storm, securing several awards and press releases, and within a year being installed at over 45 super computing centers across the globe [@Kurtzer2017-x]. 
 
+## Sharing of Containers for Reproducible Science
[…]
