### 2017-10-08 #426 @vsoch
Dear @smoe, I have done ample changes to the [paper.md](https://github.com/singularityhub/sregistry/blob/master/paper/paper.md#a-need-for-reproducible-science) and will move them (from our conversation above) here.

# Submission Changes

- [ ] I have changed the paper.md to include a better introduction for the individual that may not know about containers. I added clearly defined headers so the reader can jump ahead to a particular section of interest.
- [ ] I saw that you didn't check that there were install instructions, and so I changed the name of the page (where they are) from "dependencies" to "install" so it's more clear.
- [ ] paper.md: to add clarification for how functionality and contributors differ between singularity hub and registry, I added a new [FAQ](https://github.com/singularityhub/sregistry/blob/master/docs/faq.md) page. To this page I added some of our previous discussion as well.
- [ ] I added more references, although I maintained the .bib format. I like the RMarkdown that you linked, but other repository papers seem to use bibtex?
- [ ] I added the badge to the repository, and one more image.

@pjotrp - I don't know much about the interplanetary file system, but it sounds interesting! I will take a look. I think backup for storage is important and something we can think about and test with various users, and eventually provide a "best practice" suggestion. To foreshadow that, I added a [section to the FAQ](https://github.com/singularityhub/sregistry/blob/master/docs/faq.md#why-isnt-the-storage-backed-up)
