### 2017-10-08 c8350f6
updating paper
