+++ paper/paper.bib
@@ -205,3 +205,85 @@
   language    = "en"
 }
 
+
+@ARTICLE{Boettiger2014-cz,
+  title         = "An introduction to Docker for reproducible research, with
+                   examples from the {R} environment",
+  author        = "Boettiger, Carl",
+  abstract      = "As computational work becomes more and more integral to many
+                   aspects of scientific research, computational
+                   reproducibility has become an issue of increasing importance
+                   to computer systems researchers and domain scientists alike.
+                   Though computational reproducibility seems more straight
+                   forward than replicating physical experiments, the complex
+                   and rapidly changing nature of computer environments makes
+                   being able to reproduce and extend such work a serious
+                   challenge. In this paper, I explore common reasons that code
+                   developed for one research project cannot be successfully
+                   executed or extended by subsequent researchers. I review
+                   current approaches to these issues, including virtual
+                   machines and workflow systems, and their limitations. I then
+                   examine how the popular emerging technology Docker combines
+                   several areas from systems research - such as operating
+                   system virtualization, cross-platform portability, modular
+                   re-usable elements, versioning, and a `DevOps' philosophy,
+                   to address these challenges. I illustrate this with several
+                   examples of Docker use with a focus on the R statistical
+                   environment.",
+  month         =  oct,
+  year          =  2014,
+  archivePrefix = "arXiv",
+  primaryClass  = "cs.SE",
+  eprint        = "1410.0846"
+}
+
+@ARTICLE{Santana-Perez2015-wo,
+  title     = "Towards Reproducibility in Scientific Workflows: An
+               {Infrastructure-Based} Approach",
+  author    = "Santana-Perez, Idafen and P{\'e}rez-Hern{\'a}ndez, Mar{\'\i}a S",
+  abstract  = "It is commonly agreed that in silico scientific experiments
+               should be executable and repeatable processes. Most of the
+               current approaches for computational experiment conservation and
+               reproducibility have focused so far on two of the main
+               components of the experiment, namely, data and method. In this
+               paper, we propose a new approach that addresses the third
+               cornerstone of experimental reproducibility: the equipment. This
+               work focuses on the equipment of a computational experiment,
+               that is, the set of software and hardware components that are
+               involved in the execution of a scientific workflow. In order to
+               demonstrate the feasibility of our proposal, we describe a use
+               case scenario on the Text Analytics domain and the application
+               of our approach to it. From the original workflow, we document
+               its execution environment, by means of a set of semantic models
+               and a catalogue of resources, and generate an equivalent
+               infrastructure for reexecuting it.",
+  journal   = "Sci. Program.",
+  publisher = "Hindawi Publishing Corporation",
+  volume    =  2015,
+  month     =  feb,
+  year      =  2015
+}
+
+@ARTICLE{Wandell2015-yt,
+  title         = "Data management to support reproducible research",
+  author        = "Wandell, B A and Rokem, A and Perry, L M and Schaefer, G and
+                   Dougherty, R F",
+  abstract      = "We describe the current state and future plans for a set of
+                   tools for scientific data management (SDM) designed to
+                   support scientific transparency and reproducible research.
+                   SDM has been in active use at our MRI Center for more than
+                   two years. We designed the system to be used from the
+                   beginning of a research project, which contrasts with
+                   conventional end-state databases that accept data as a
+                   project concludes. A number of benefits accrue from using
+                   scientific data management tools early and throughout the
+                   project, including data integrity as well as reuse of the
+                   data and of computational methods.",
+  month         =  feb,
+  year          =  2015,
+  archivePrefix = "arXiv",
+  primaryClass  = "q-bio.QM",
+  eprint        = "1502.06900"
+}
+
+
