### 2017-10-08 #426 @vsoch
@pjotrp I shorted the introduction quite a bit, and moved the bulk of the background into the docs [introduction](https://singularityhub.github.io/sregistry/introduction.html). [Here is the updated paper](https://github.com/singularityhub/sregistry/blob/master/paper/paper.md). Your suggested introduction is fantastic, and I only tweaked it slightly (Docker in fact has huge security issues so it was a non starter for HPC).

I left the remainder of the description with pictures assuming that interested readers will continue, and dis-interested will have been satisfied by the early description and move on.
