+++ paper/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 Singularity Registry is a non-centralized free and Open Source infrastructure to facilitate management and sharing of institutional or personal containers.
 
-A container is the encapsulation of an entire computational environment that can be run consistently if the platform supports it. It is an aid in reproducibility [@ARTICLE{Moreews2015-dy;@ARTICLE{Belmann2015-eb;@ARTICLE{Boettiger2014-cz;@ARTICLE{Santana-Perez2015-wo;@ARTICLE{Wandell2015-yt] because different researchers can run the exact same software stack on different underlying (Linux Intel) systems. Docker [@ARTICLE{Merkel2014-da] has become popular as a general container system because it allows software to be bundled with administrator privileges, however,  it poses great security risks if installed on a multi-user, shared computational resource, and thus Docker has not had admission to these high performance computing (HPC) environments. Singularity [@Kurtzer2017-x] offers similar features to Docker for software deployment but does not not allow the user to escalate to root, and so it has been easy to accept in HPC environments since its introduction. So far, Singularity has been deployed in over 50 HPC resources across the globe.
+A container is the encapsulation of an entire computational environment that can be run consistently if the platform supports it. It is an aid in reproducibility [@ARTICLE{Moreews2015-dy;@ARTICLE{Belmann2015-eb;@ARTICLE{Boettiger2014-cz;@ARTICLE{Santana-Perez2015-wo;@ARTICLE{Wandell2015-yt] because different researchers can run the exact same software stack on different underlying (Linux Intel) systems. Docker [@ARTICLE{Merkel2014-da] has become popular as a general container system because it allows software to be bundled with administrator privileges, however,  it poses great security risks if installed on a multi-user, shared computational resource, and thus Docker has not had admission to these high performance computing (HPC) environments. Singularity [@Kurtzer2017-x] offers similar features to Docker for software deployment but does not not allow the user to escalate to root, and so it has been easy to accept in HPC environments. Since its introduction, Singularity has been deployed in over 50 HPC resources across the globe.
 
 ## Sharing of Containers for Reproducible Science
 Essential to the success of Singularity is not just creation of images, but sharing of them. To address this need, a free cloud service, Singularity Hub [@SingularityHub], was developed to build and share containers for scientists simply by way of building containers from a specification file in a version controlled Github repository. This setup is ideal given a small number of containers, each belonging in one Github repository, but is not optimal for institutions that wanted to build at scale using custom strategies.
