### 2017-10-08 #426 @vsoch
@pjotrp that is correct - inside a Singularity container you are the same user as outside the container:

```
$ whoami
vanessa
$ singularity shell docker://ubuntu
Docker image path: index.docker.io/library/ubuntu:latest
Cache folder set to /home/vanessa/.singularity/docker
Creating container runtime...
Singularity: Invoking an interactive shell within container...

Singularity ubuntu:~/Documents/Dropbox/Code/sregistry/sregistry> whoami
vanessa
```

And thus if you run as root, you are root inside as well!

```
$ sudo whoami
root
sudo singularity shell docker://ubuntu
Docker image path: index.docker.io/library/ubuntu:latest
Cache folder set to /root/.singularity/docker
[5/5] |===================================| 100.0%
Creating container runtime...
Singularity: Invoking an interactive shell within container...

Singularity ubuntu:/home/vanessa/Documents/Dropbox/Code/sregistry/sregistry> whoami
root
```
