### 2017-10-08 #426 @pjotrp
OK, I think that means someone needs to be root to create containers? That would be an HPC administrator? How would apt or rpm work otherwise?
