### 2017-10-08 #426 @vsoch
@pboesu yes, the typical workflow is to build on a resource (a personal work station,for example) with root, and then run the container anywhere. The user running the container cannot escalate to root on the host system, as would be possible with Docker. It matches nicely to a typical scientific workflow, where you write and test scripts on a personal computer, and then run at scale. The workflow typically look like this (with Singularity 2.3 and lower)

![image](https://user-images.githubusercontent.com/814322/31320031-9e316f04-ac22-11e7-9cac-6ae25d7bf4aa.png)

For a hosted registry, the three general use cases I have in mind are the following:

 - a personal registry on a laptop, personal computer, the user is root and uses Singularity Registry to easily manage and run containers (this is how I use it).
- an institutional registry where administrators build the containers, and push for users to access. This is akin to cluster managers installing and managing software modules,etc.
 - A collaborative institutional registry where the administrators manage one or more repositories of build recipes, and users are allowed to contribute with PRs to images that will be built. The containers and changes are discussed on PR and issue boards, and merge only done when everyone is happy. The building /pushing itself could happen on continuous integration (after tests pass) or some build server (meaning the administrator would pull from Github, build, and push to the registry).

For using, meaning running, executing, shelling (all the functions important for working on a cluster) the user's have the same permissions as they would as their normal user. Binding the host to the container is done automatically for some folders (eg, tmp, the present working directory, the user's home) and on request with `--bind`.
