### 2017-10-09 0beff3e
adding use cases and more faq
