### 2017-10-09 #426 @vsoch
Thanks @pjotrp, I don't in fact think the use cases were properly reflected, and they are important to give users a sense of "who cares?" I've added a [use cases](https://singularityhub.github.io/sregistry/use-cases.html) section to the docs to address this.
