### 2017-10-09 singularityhub/sregistry#35 @smoe
Promote/create Debian/Ubuntu packages for installation
I found https://github.com/singularityhub/sregistry/blob/master/scripts/prepare_instance.sh to be an interesting mix between Ubuntu specifica and general install instructions with pip and the downloaded Anaconda. Is Anaconda truly required or could we also just work with a regular installable (and deinstallable (!)) system-wide installed python?

The two pip-installed packages do no require the extra pip, for instance https://packages.debian.org/search?keywords=python-ipaddress or https://packages.debian.org/search?keywords=python-oauth2client, so these packages I also expect in Ubuntu. Does this possibly mean that pip does not need to be installed, either?

I can update or backport for you as required.
