### 2017-10-09 #426 @pjotrp
@smoe could you try a quick install to be able to tick the *Functionality* boxes? It is a requirement rather. No rush, you have been quick so far.
