### 2017-10-09 singularityhub/sregistry#35 @vsoch
hey @smoe ! Anaconda isn't necessary, it's just the distribution of python I tend to go for in the case of needing nicely packaged numerical libraries. The basic dependencies are:

 - docker-compose (needs the python libraries on the host)
 - git to pull repo
 - docker

And it doesn't matter so much how you choose to install, or the host that is your preference. I included the prepare instance script (the one that I use) if others were interested, the user might already have these things installed (eg system python) and just need to install Docker.

pip as a package manager is always my preference for installing, but to each his own! :)
