### 2017-10-09 92392bd
adding dois to references for which they exist!
