+++ paper/paper.bib
@@ -17,6 +17,7 @@
                previously unmet need, makes Singularity a game changing
                development for computational science.",
   journal   = "PLoS One",
+  doi       = "journal.pone.0177459",
   publisher = "Public Library of Science",
   volume    =  12,
   number    =  5,
@@ -67,7 +68,7 @@
 }
 
 
-@ARTICLE{Merkel2014-da,
+@MISC{Merkel2014-da,
   title     = "Docker: Lightweight Linux Containers for Consistent Development
                and Deployment",
   author    = "Merkel, Dirk",
@@ -105,6 +106,7 @@
   journal   = "Conference on Bioinformatics and …",
   publisher = "Springer International Publishing",
   pages     = "587--601",
+  doi       = "https://doi.org/10.1007/978-3-319-31744-1_52"
   month     =  apr,
   year      =  2016,
   language  = "en"
@@ -166,7 +168,8 @@
   year        =  2015,
   keywords    = "bioinformatics; community driven registry; container;
                  deployment; docker; interoperability; maintainability",
-  language    = "en"
+  language    = "en",
+  doi         = "10.12688/f1000research.7536.1"
 }
 
 @ARTICLE{Belmann2015-eb,
@@ -200,6 +203,7 @@
   pages       = "47",
   month       =  oct,
   year        =  2015,
+  doi         = "10.1186/s13742-015-0087-0",
   keywords    = "Bioinformatics; Docker; Reproducibility; Software; Standards;
                  Usability",
   language    = "en"
@@ -234,6 +238,7 @@
   year          =  2014,
   archivePrefix = "arXiv",
   primaryClass  = "cs.SE",
+  doi           = "10.1145/2723872.2723882",
   eprint        = "1410.0846"
 }
 
@@ -261,7 +266,8 @@
   publisher = "Hindawi Publishing Corporation",
   volume    =  2015,
   month     =  feb,
-  year      =  2015
+  year      =  2015,
+  doi       = "http://dx.doi.org/10.1155/2015/243180"
 }
 
 @ARTICLE{Wandell2015-yt,
@@ -285,5 +291,3 @@
   primaryClass  = "q-bio.QM",
   eprint        = "1502.06900"
 }
-
-
