### 2017-10-09 #426 @smoe
Sigh. How comes I knew this would happen. I ran what I understood to be run and ran into an issue. @vsoch, I get you an account on that machine (regular Debian Stretch) if that is of any help. I followed the instructions on https://github.com/singularityhub/sregistry/blob/master/scripts/prepare_instance.sh with docker already installed.

I successfully run the hello-world as root. As a regular user I get a "permission denied" for building in the fresh sregistry git checkout. I am not fluent enough with docker - should the
"docker build ." possibly also be executed with sudo and that was only working on your site without it since you already downloaded everything you need to download? Over here it is quite busy to build images when I first initiated the build.

So, running with sudo, I end up with:

```
.... many lines omitted ....
update-alternatives: using /usr/bin/vim.basic to provide /usr/bin/rview (rview) in auto mode
update-alternatives: using /usr/bin/vim.basic to provide /usr/bin/vi (vi) in auto mode
update-alternatives: using /usr/bin/vim.basic to provide /usr/bin/view (view) in auto mode
update-alternatives: using /usr/bin/vim.basic to provide /usr/bin/ex (ex) in auto mode
update-alternatives: using /usr/bin/vim.basic to provide /usr/bin/editor (editor) in auto mode
Setting up libtasn1-doc (4.2-3+deb8u3) ...
Processing triggers for systemd (215-17+deb8u4) ...
Processing triggers for libc-bin (2.19-18+deb8u4) ...
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libMagickCore-6.Q16.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libMagickWand-6.Q16.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libcairo-gobject.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libcairo-script-interpreter.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libcairo.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libgdk_pixbuf-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libgdk_pixbuf_xlib-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libgio-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libglib-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libgmodule-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libgobject-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libgthread-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/librsvg-2.so: No such file or directory
 ---> e320dd67f67f
Removing intermediate container 9d4fc33d83e1
Step 4/21 : RUN git clone https://www.github.com/singularityware/singularity.git
 ---> Running in acfb57909644
Cloning into 'singularity'...
 ---> 4fa9bdac1522
Removing intermediate container acfb57909644
Step 5/21 : WORKDIR singularity
 ---> 4dd86eced674
Removing intermediate container c81613b8fbe6
Step 6/21 : RUN ./autogen.sh && ./configure --prefix=/usr/local && make && make install
 ---> Running in eb2b8cd2c10b
+autoreconf -i -f
aclocal: error: aclocal: file '/usr/share/aclocal/gsettings.m4' does not exist
autoreconf: aclocal failed with exit status: 1
The command '/bin/sh -c ./autogen.sh && ./configure --prefix=/usr/local && make && make install' returned a non-zero code: 1
```

If I get this right, then this was all executed in the docker environment, so it should not have mattered that I did not have the autoconf etc. package installed. The missing gsettings.m4 is in the libglic2.0-dev package:

```
$ apt-file search gsettings.m4
gnome-builder: /usr/share/gnome-builder/plugins/autotools_templates/resources/m4/gsettings.m4
libglib2.0-dev: /usr/share/aclocal/gsettings.m4
```

For the better overview of yours, starting the process again, I get:

```
moeller@server2:~/sregistry$ docker build  .
Got permission denied while trying to connect to the Docker daemon socket at unix:///var/run/docker.sock: Post http://%2Fvar%2Frun%2Fdocker.sock/v1.32/build?buildargs=%7B%7D&cachefrom=%5B%5D&cgroupparent=&cpuperiod=0&cpuquota=0&cpusetcpus=&cpusetmems=&cpushares=0&dockerfile=Dockerfile&labels=%7B%7D&memory=0&memswap=0&networkmode=default&rm=1&session=d35d88477cc93e251cb5dae3c5fbf624c49cccdf9a0aaf06c4875279daeb60d3&shmsize=0&target=&ulimits=null: dial unix /var/run/docker.sock: connect: permission denied
moeller@server2:~/sregistry$ ^Ccker build  .
moeller@server2:~/sregistry$ sudo docker build -t vanessa/sregistry .
Sending build context to Docker daemon  22.77MB
Step 1/21 : FROM python:3.5.1
 ---> a00e9008965a
Step 2/21 : ENV PYTHONUNBUFFERED 1
 ---> Using cache
 ---> 41200125058b
Step 3/21 : RUN apt-get update && apt-get install -y     anacron     libopenblas-dev     gfortran     pkg-config     libxml2-dev     libxmlsec1-dev     libhdf5-dev     libgeos-dev     build-essential     openssl     wget     vim
 ---> Using cache
 ---> e320dd67f67f
Step 4/21 : RUN git clone https://www.github.com/singularityware/singularity.git
 ---> Using cache
 ---> 4fa9bdac1522
Step 5/21 : WORKDIR singularity
 ---> Using cache
 ---> 4dd86eced674
Step 6/21 : RUN ./autogen.sh && ./configure --prefix=/usr/local && make && make install
 ---> Running in c4bc5a22558f
+autoreconf -i -f
aclocal: error: aclocal: file '/usr/share/aclocal/gsettings.m4' does not exist
autoreconf: aclocal failed with exit status: 1
The command '/bin/sh -c ./autogen.sh && ./configure --prefix=/usr/local && make && make install' returned a non-zero code: 1
```

On a sidenote, the python3.5.1 image you ask for is based on Debian Jessie. That is the oldstable one - maintained for another while, but ... it should be the current release "Stretch", not "Jessie".  I just checked the images for docker and among the readily prepared images it seems like that python 3.6.3 is the closest choice that uses Stretch (https://store.docker.com/images/python). Frankly, though, so much gets installed for that image, that it does not really matter if there is a Python already shipping or not. I propose you just get an empty Debian image and add python 3.5 with "apt-get install python3" - as 3.5 is still the default python3 installation (https://packages.qa.debian.org/p/python3-defaults.html).

So, after taking some deep breath I propose we just get this fixed. My offer to get you an account on the machine still holds (it is not doing anything but helping Einstein@Home these days, so don't worry and send me a public ssh key if you like).
