### 2017-10-09 #426 @vsoch
Hey @smoe ! Sorry for your troubles! There are two things here - preparing your host to use Docker and compose, and then running the image. I

## Preparing Host
It's best to follow the instructions for your distribution [here](https://docs.docker.com/engine/installation/). Note that I provided instructions for an Ubuntu host. There are detailed instructions provided by Docker for non ubuntu, along with [docker-compose](https://docs.docker.com/compose/install/).

** Important ** you need to, after you add your user to Docker's group, log out and back in before you continue. The call to `docker run hello-world` should succeed.

## Bringing up Registry
You shouldn't need to build the image, but if you want to:

```
docker build -t vanessa/sregistry .
```
Note that I didn't run just `docker build .` as you did.  However, the entire image is provided pre-build on Docker Hub:

```
https://hub.docker.com/r/vanessa/sregistry/builds/
```
so you should be able to do:

```
docker-compose up -d
```

Let's walk through these steps, one at a time, and discuss when you hit a snag (and please don't move on). I've had two others deploy without much issue, so I think it comes down to your installation of docker and compose. I will update the docs to explicitly say that the prepare script example that I provided is for an ubuntu host.
