### 2017-10-09 singularityhub/sregistry#35 @smoe
Thank you for your swift reply, @vsoch. I had very much hoped you would say something along those lines - since this means that I have done everything right-enough for my test run.

Newbies like me just don't know about version dependencies and may redundantly install bits and pieces. Please allow that I prepare a pull request that would then be minimalistic and leave all docker-specific installation to ... well docker: The requirements docker needs will be auto-fetched from the repository (been there a few days ago). This means that you can remove all installation-related lines of python/python-modules/pip from that script. Your software then looks even more simple to install and better understandable for users of other distributions.

Please allow that I eventually come up with a pull request with something that is both minimalistic and working for me.
