### 2017-10-09 25b7701
making docs more specific to install for distribution and log in and out
