### 2017-10-09 #426 @vsoch
hey @smoe ! I updated the docs to very specifically say to install docker an docker-compose for your particular choice, along with needing to log in and out.

If you want to see one of my colleagues deployed sregistry, check out [randomroad](https://sregistry.randomroad.net)

Please let me know what is (and isn't clear) and I'll improve the instructions in the docs. I want to make sure they are user friendly for a wide range of audiences.
