### 2017-10-09 singularityhub/sregistry#35 @vsoch
hey @smoe ! I think I just pushed change that address your concern - I point the user to the Docker and Docker compose installation pages for instructions specific to the distribution. When you are willing to dive in again, please try the install and bringing up the image (and you could skip building and it will pull the image from docker hub). Let me know what needs work (and a pull request with suggestions would be fantastic!) But definitely I can do most of the heavy lifting to fit the issues that you find.

If you need to remove some previously built erroneous image, try:

```
docker stop $(docker ps -a -q)
docker rm $(docker ps -a -q)
```

It's great you are learning some too! Docker is... a bit weird. It took me definitely some time to get the hang of it.
