### 2017-10-09 singularityhub/sregistry#35 @smoe
I cannot see it - have you pushed?
