### 2017-10-09 singularityhub/sregistry#35 @smoe
Found it https://github.com/singularityhub/sregistry/blob/master/docs/install.md - reads very nice, indeed! Many thanks!
