### 2017-10-09 #426 @smoe
I had used the wrong open shell for the build in which I had just added me to the docker user group. So, thank you for that.  Indeed, I can now run as a regular user for which I had to use sudo before. Good news first: The docker-compose just worked fine. For @pjotrp:

```
moeller@server2:~/sregistry$ docker-compose up -d
Pulling redis (redis:latest)...
latest: Pulling from library/redis
065132d9f705: Pull complete
be9835c27852: Pull complete
f4a0d1212c38: Pull complete
43be9e9f0fb9: Pull complete
a1bca8e532ec: Pull complete
382eae952932: Pull complete
Digest: sha256:ebb396dc3ac00e8eb4a64c1c022ef41ef16801f31ff98b16916a77fdc7252e67
Status: Downloaded newer image for redis:latest
Pulling db (postgres:latest)...
latest: Pulling from library/postgres
219d2e45b4af: Pull complete
87b4d6274d77: Pull complete
2569a32ee6dd: Pull complete
23b4d0fc3192: Pull complete
8275aae461c7: Pull complete
45087ee6fc31: Pull complete
2d2265f720a6: Pull complete
2860cdb4889c: Pull complete
597205e436ef: Pull complete
d445130f222d: Pull complete
def74255ce78: Pull complete
4a3042ea21df: Pull complete
686ac5adc42e: Pull complete
Digest: sha256:242a4569b89cda783c164380aaa202dc0332d79c49b856a0554c3a94de8a4dfa
Status: Downloaded newer image for postgres:latest
Pulling uwsgi (vanessa/sregistry:latest)...
latest: Pulling from vanessa/sregistry
51f5c6a04d83: Already exists
65e9ddd8bd7a: Already exists
c41545ebedf5: Already exists
04aed1875617: Already exists
cf8af6bdf113: Already exists
b1dde6937fab: Already exists
c8bc9f75687c: Already exists
a91d9d9365df: Pull complete
99c405b8c55e: Pull complete
0e4bb0a0976e: Pull complete
b225dcdce7e1: Pull complete
ab1b2e6030c9: Pull complete
9ac9ee44eb32: Pull complete
7c87a686bfaa: Pull complete
fafbd6f79b25: Pull complete
1006ffb0701c: Pull complete
a065aeb43100: Pull complete
bbe9db7595f1: Pull complete
af848c89cfcb: Pull complete
ec8d609a6d9b: Pull complete
d0110b169e20: Pull complete
Digest: sha256:57be162740e34e8bd415b957c137a9b98116c5b23d2be36ec8947d2602f74dde
Status: Downloaded newer image for vanessa/sregistry:latest
Pulling nginx (nginx:latest)...
latest: Pulling from library/nginx
afeb2bfd31c0: Pull complete
7ff5d10493db: Pull complete
d2562f1ae1d0: Pull complete
Digest: sha256:af32e714a9cc3157157374e68c818b05ebe9e0737aac06b55a09da374209a8f9
Status: Downloaded newer image for nginx:latest
Creating sregistry_redis_1
Creating sregistry_db_1
Creating sregistry_uwsgi_1
Creating sregistry_nginx_1
Creating sregistry_worker_1
```
Executed with sudo or after properly being accepted as a member of the docker group on the system, the error remains the same:

```
moeller@server2:~/sregistry$ docker build -t vanessa/sregistry .
Sending build context to Docker daemon  22.79MB
Step 1/21 : FROM python:3.5.1
 ---> a00e9008965a
Step 2/21 : ENV PYTHONUNBUFFERED 1
 ---> Using cache
 ---> 41200125058b
Step 3/21 : RUN apt-get update && apt-get install -y     anacron     libopenblas-dev     gfortran     pkg-config     libxml2-dev     libxmlsec1-dev     libhdf5-dev     libgeos-dev     build-essential     openssl     wget     vim
 ---> Using cache
 ---> e320dd67f67f
Step 4/21 : RUN git clone https://www.github.com/singularityware/singularity.git
 ---> Using cache
 ---> 4fa9bdac1522
Step 5/21 : WORKDIR singularity
 ---> Using cache
 ---> 4dd86eced674
Step 6/21 : RUN ./autogen.sh && ./configure --prefix=/usr/local && make && make install
 ---> Running in 849c7875abcc
+autoreconf -i -f
aclocal: error: aclocal: file '/usr/share/aclocal/gsettings.m4' does not exist
autoreconf: aclocal failed with exit status: 1
The command '/bin/sh -c ./autogen.sh && ./configure --prefix=/usr/local && make && make install' returned a non-zero code: 1
```

Does the build still work for you? Concerning the documentation, I just followed https://github.com/singularityhub/sregistry/blob/master/scripts/prepare_instance.sh to which your documentation links and that does the build. But then again, our dear editor wants me to build it, so, the pointer to prepare_instance.sh was fine to this point. I had not even seen install.md before your kind indirect pointer on https://github.com/singularityhub/sregistry/issues/35 .

You are right in that you are granting me the opportunity to learn some docker over the process. Hey, I had to install it when what I really wanted to install was Singularity :)  Please allow that I continue with my learning tomorrow, night time over here. Many thanks for all your friendly and quick help.
