### 2017-10-09 52155bd
updating singularity install to use development
