+++ Dockerfile
@@ -5,6 +5,12 @@ RUN apt-get update && apt-get install -y \
     libopenblas-dev \
     gfortran \
     pkg-config \
+    libtool \
+    autotools-dev \
+    squashfs-tools \
+    debootstrap \
+    automake \
+    autoconf \
     libxml2-dev \
     libxmlsec1-dev \
     libhdf5-dev \
@@ -15,7 +21,7 @@ RUN apt-get update && apt-get install -y \
     vim
 
 # Install Singularity
-RUN git clone https://www.github.com/singularityware/singularity.git
+RUN git clone -b development https://www.github.com/singularityware/singularity.git
 WORKDIR singularity
 RUN ./autogen.sh && ./configure --prefix=/usr/local && make && make install
 
