### 2017-10-09 #426 @vsoch
It looks like it's having trouble with singularity itself (and not the registry). I updated the image to clone from development, if you want to try building again. Here is also a complete build I did from scratch to confirm that it works for me: https://asciinema.org/a/141559?speed=3

Do you think it's too easy to find the prepare_instance.sh script (over the install instructions?) If so, let's definitely fix that.

No worries, chat tomorrow! Do
