### 2017-10-10 c216856
adding default pagination class to settings
