### 2017-10-10 f3cf05a
adding api authentication notes
