### 2017-10-10 587373d
modified:   docs/credentials.md
