### 2017-10-10 81caa1e
Dockerfile adds gsettings.m4, fixes build error
The only idea coming to mind why on this end the build fails but not for @vsoch and friends are differences between the dynamically assigned Debian mirrors that depend on where one is located. That line fixed it (except that I had it on a separate RUN apt-get ... invocation).
