### 2017-10-10 #426 @smoe
Sigh. I have removed all images (and running containers), and even after your change to the Dockerfile, I still get
```
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libglib-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libgmodule-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libgobject-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/libgthread-2.0.so: No such file or directory
/sbin/ldconfig.real: Cannot lstat /usr/lib/x86_64-linux-gnu/librsvg-2.so: No such file or directory
 ---> dbba45aef6b8
Removing intermediate container cbc044f24d7c
Step 4/21 : RUN git clone https://www.github.com/singularityware/singularity.git
 ---> Running in fdcc79b73821
Cloning into 'singularity'...
 ---> dd68929297c4
Removing intermediate container fdcc79b73821
Step 5/21 : WORKDIR singularity
 ---> 2087f0109a29
Removing intermediate container 137c1d022d4d
Step 6/21 : RUN ./autogen.sh && ./configure --prefix=/usr/local && make && make install
 ---> Running in c2bda8ebff01
+autoreconf -i -f
aclocal: error: aclocal: file '/usr/share/aclocal/gsettings.m4' does not exist
autoreconf: aclocal failed with exit status: 1
The command '/bin/sh -c ./autogen.sh && ./configure --prefix=/usr/local && make && make install' returned a non-zero code: 1
```
but (and this is a big but) https://github.com/singularityhub/sregistry/pull/39 fixed that for me. In my commit note I speculated that the mirrors of Debian may differ slightly in their punctuality of updates. Or do you have a web cache for the pip installation?

Anyway, I can proceed to the configuration now.
