### 2017-10-10 #426 @vsoch
Your fix is added! I think it might also be missing cmake? It could be a different version related to mirrors. I've merged your additional library, and also just added cmake. It's very strange, because it works for me (and the few others that have installed it) so hard to pin down what it might be. I also read that the error pops up when package config fails to link the file.
