### 2017-10-10 #426 @smoe
Ok ... is there a place for me to look when encounering an "internal server error" that is displayed as the only content? To be seen on server2.ffoh.de.
