### 2017-10-10 #426 @smoe
A missing cmake was not complained about.

Is there a chance to  have you migrate to Debian Stretch? I volunteer to perform the test build.
