### 2017-10-10 #426 @vsoch
Try

```
docker-compose logs uwsgi
```
and please tell me what you find!
