### 2017-10-10 #426 @vsoch
You shouldn't need to worry about building on your particular host. I can't explain the issues that you are having, but since there is a pre-built container provided --> https://hub.docker.com/r/vanessa/sregistry/builds/ that should be more than sufficient to reproduce the application exactly as intended. Building on your own, actually, would produce a different image (as we have seen) given that different mirrors / other versions could be obtained with installation commands.
