### 2017-10-10 #426 @vsoch
I think it may be the case that you need to follow the instructions in setup first **before** you start the image. Did you see the page here?

https://singularityhub.github.io/sregistry/install.html

That configuration needs to be done before you start the image.  If you are using the one from docker hub, the files are mounted so they will be updated when you bring up the image.
