### 2017-10-10 #426 @vsoch
Hi @smoe - please see my instruction above:

```
docker-compose logs uwsgi
```
that will show the logs of the main server, which is the image called uwsgi.  You do need to follow the instructions and choose some method to authenticate. The other option is that I could allow you (the user) to bring up a registry that you are unable to login to.  Please run that command and show me the output (which will correspond to the server error). Minimally we can set up dummy variables so the server comes up without error, albeit you probably won't be able to do much without the configurations.
