### 2017-10-11 45a429b
fixing bug with commands page
