### 2017-10-11 #426 @smoe
It says that the SECRET_KEY would be empty - which it is not on my end. I also failed to see at which point you forward that file into your containers - but I have complete trust that we can settle that independently from this review. So I pressed the last button.

I personally like the idea of a "display only" image server very much where users could anonymously download the images they want. This could possibly come very handy e.g. for the CWL community to automatically prepare images for whatever favorite pipelines of theirs. Feedback on those images could be with some github solution elsewhere.

I had another look at your infrastructure and you indeed set up everything with python3 on Debian Jessie.  The prior noted location dependency is presumably caused by the automated mirror finding with httpredir. I find this somewhat shocking, I must admit.

```
~/sregistry$ docker run -it  python:3.5.1 bash
root@717f5c53a176:/# pip --version
pip 8.1.2 from /usr/local/lib/python3.5/site-packages (python 3.5)
root@717f5c53a176:/# cat /etc/apt/sources.list
deb http://httpredir.debian.org/debian jessie main
deb http://httpredir.debian.org/debian jessie-updates main
deb http://security.debian.org jessie/updates main
```

To help your new users it would be nice if you could catch the one or other error and come up with a minimal system.

```
uwsgi_1   | Wed Oct 11 11:48:47 2017 - *** Operational MODE: single process ***
uwsgi_1   | Traceback (most recent call last):
uwsgi_1   |   File "shub/wsgi.py", line 7, in <module>
uwsgi_1   |     application = get_wsgi_application()
uwsgi_1   |   File "/usr/local/lib/python2.7/dist-packages/django/core/wsgi.py", line 13, in get_wsgi_application
uwsgi_1   |     django.setup(set_prefix=False)
uwsgi_1   |   File "/usr/local/lib/python2.7/dist-packages/django/__init__.py", line 22, in setup
uwsgi_1   |     configure_logging(settings.LOGGING_CONFIG, settings.LOGGING)
uwsgi_1   |   File "/usr/local/lib/python2.7/dist-packages/django/conf/__init__.py", line 56, in __getattr__
uwsgi_1   |     self._setup(name)
uwsgi_1   |   File "/usr/local/lib/python2.7/dist-packages/django/conf/__init__.py", line 41, in _setup
uwsgi_1   |     self._wrapped = Settings(settings_module)
uwsgi_1   |   File "/usr/local/lib/python2.7/dist-packages/django/conf/__init__.py", line 110, in __init__
uwsgi_1   |     mod = importlib.import_module(self.SETTINGS_MODULE)
uwsgi_1   |   File "/usr/lib/python2.7/importlib/__init__.py", line 37, in import_module
uwsgi_1   |     __import__(name)
uwsgi_1   |   File "./shub/__init__.py", line 2, in <module>
uwsgi_1   |     from .celery import shubcelery as celery_app
uwsgi_1   |   File "./shub/celery.py", line 45, in <module>
uwsgi_1   |     from opbeat.contrib.django.models import (
uwsgi_1   |   File "/usr/local/lib/python2.7/dist-packages/opbeat/contrib/django/models.py", line 226, in <module>
uwsgi_1   |     if 'opbeat.contrib.django' in django_settings.INSTALLED_APPS:
uwsgi_1   |   File "/usr/local/lib/python2.7/dist-packages/django/conf/__init__.py", line 56, in __getattr__
uwsgi_1   |     self._setup(name)
uwsgi_1   |   File "/usr/local/lib/python2.7/dist-packages/django/conf/__init__.py", line 41, in _setup
uwsgi_1   |     self._wrapped = Settings(settings_module)
uwsgi_1   |   File "/usr/local/lib/python2.7/dist-packages/django/conf/__init__.py", line 129, in __init__
uwsgi_1   |     raise ImproperlyConfigured("The SECRET_KEY setting must not be empty.")
uwsgi_1   | django.core.exceptions.ImproperlyConfigured: The SECRET_KEY setting must not be empty.
uwsgi_1   | Wed Oct 11 11:48:48 2017 - unable to load app 0 (mountpoint='') (callable not found or import error)
uwsgi_1   | Wed Oct 11 11:48:48 2017 - *** no app loaded. going in full dynamic mode ***
uwsgi_1   | Wed Oct 11 11:48:48 2017 - *** uWSGI is running in multiple interpreter mode ***
uwsgi_1   | Wed Oct 11 11:48:48 2017 - spawned uWSGI master process (pid: 48)
uwsgi_1   | Wed Oct 11 11:48:48 2017 - spawned uWSGI worker 1 (pid: 52, cores: 1)
uwsgi_1   | Wed Oct 11 11:48:55 2017 - --- no python application found, check your startup logs for errors ---
```
