### 2017-10-11 #426 @vsoch
hey @smoe! The secret key is a [django application detail](https://docs.djangoproject.com/en/1.11/ref/settings/#std:setting-SECRET_KEY), and you must have missed where I wrote about it in the [instructions to generate one](https://singularityhub.github.io/sregistry/install.html#secrets). It's important for security features for any django application, and as you discovered, it will get angry at you if you don't have it. Would it be useful to step back and go back to the instructions and go through them to make sure you set everything up proper?

>> I personally like the idea of a "display only" image server very much where users could anonymously download the images they want.

Yes! And actually a registry can serve as exactly this. This doesn't do away with the need of someone (some admin) setting it up and managing the images.  And we can do slightly better because we don't only download anonymously, we can track the number of uses (anonymously) to better understand which software is most useful.

>> This could possibly come very handy e.g. for the CWL community to automatically prepare images for whatever favorite pipelines of theirs. Feedback on those images could be with some github solution elsewhere.

Yes I had this in mind when I created push - images could be pushed automatically after a pull request (PR) and discussion on a version controlled base like Github.

>> I had another look at your infrastructure and you indeed set up everything with python3 on Debian Jessie. The prior noted location dependency is presumably caused by the automated mirror finding with httpredir. I find this somewhat shocking, I must admit.

I am using the Docker provided python image, which likely uses those mirrors. I don't have much opinion, but if you want to avoid building all together I can suggest just pulling from Docker Hub.
