### 2017-10-11 #426 @smoe
Hi @vsoch, admittedly, my very initial run was without the secret key since the prepare_instance.sh script did not generate it. But a next invocation, with all images removed if I did not mess it up, had it in place. And this is how it still is:
```
moeller@server2:~/sregistry$ cat shub/settings/secret.py
SECRET_KEY='djangoSecretKeytyw!b^d#elqz2d@c!m)o)g=pfnkilc07mof=pa8f@zszl5^fe=forSregistry'
moeller@server2:~/sregistry$ ls -l shub/settings/secret.py
-rw-r--r-- 1 moeller moeller 91 Oct 10 17:43 shub/settings/secret.py
```
I had used the django generator you referred to create it.

Concerning the python image I am still a bit unhappy. It seems like this comes with the autotools already installed, and it even has git.  Why does have python image feature all that?

If you do not mind too much, please add those bits in - "back in" for autoconf and friends.
    git \
    autoconf \
    automake \
    libtool

This is just because it is needed and might not work once you upgrade to a different version of that image. Btw, I had a building version with Debian Stretch instead of Debian Jessie-based python:3.5.1.
