### 2017-10-11 c996181
updating docker image and adding additional docs to clarify restart
