+++ Dockerfile
@@ -1,20 +1,27 @@
 FROM python:3.5.1
 ENV PYTHONUNBUFFERED 1
 RUN apt-get update && apt-get install -y \
+    pkg-config \
+    cmake \
+    openssl \
+    wget \
+    git \
+    vim
+
+RUN apt-get update && apt-get install -y \
     anacron \
+    autoconf \
+    automake \
+    libtool \
     libopenblas-dev \
     libglib2.0-dev \
     gfortran \
-    pkg-config \
     libxml2-dev \
     libxmlsec1-dev \
     libhdf5-dev \
     libgeos-dev \
-    build-essential \
-    cmake \
-    openssl \
-    wget \
-    vim
+    build-essential
+
 
 # Install Singularity
 RUN git clone https://www.github.com/singularityware/singularity.git
