### 2017-10-11 #426 @vsoch
hey @smoe! Are you still have trouble with the secret key? You shouldn't need to remove or rebuild images, just restart with docker compose (and this is true whenever a change is done and the images are running):

```
docker-compose restart
```
I added more verbose documentation about this to [this section](https://singularityhub.github.io/sregistry/install.html#compose-the-images), in case you missed it.

I can't comment on why the base python chooses to include core git (likely a choice of the developer) but per your request, I've added back in the libraries you listed.
