### 2017-10-11 #426 @smoe
```
moeller@server2:~/sregistry$ docker rmi vanessa/sregistry
Error: No such image: vanessa/sregistry
moeller@server2:~/sregistry$ docker-compose up -d
Pulling uwsgi (vanessa/sregistry:latest)...
latest: Pulling from vanessa/sregistry
51f5c6a04d83: Already exists
65e9ddd8bd7a: Already exists
c41545ebedf5: Already exists
04aed1875617: Already exists
cf8af6bdf113: Already exists
b1dde6937fab: Already exists
c8bc9f75687c: Already exists
7ec2ccb4b7ef: Pull complete
45aa57212348: Pull complete
bbe3ffa886f4: Pull complete
ae4988fdcaba: Pull complete
88774b19f593: Pull complete
a9fbd563a04b: Pull complete
d66e99c8cb84: Pull complete
03c3b0adfe39: Pull complete
4e7c6ac411f1: Pull complete
d2ffd7d5720b: Pull complete
0ed82f0bc059: Pull complete
7a302558f857: Pull complete
af4abfaab25d: Pull complete
e9347ba1dfcd: Pull complete
57c07257b2a4: Pull complete
Digest: sha256:12d7c6ddc801f07813dd068308048b76637488255d3219bb7e91906531b35e87
Status: Downloaded newer image for vanessa/sregistry:latest
Creating sregistry_redis_1
Creating sregistry_db_1
Creating sregistry_uwsgi_1
Creating sregistry_nginx_1
Creating sregistry_worker_1
moeller@server2:~/sregistry$ pwd
/home/moeller/sregistry
moeller@server2:~/sregistry$ cat shub/settings/secret.py
SECRET_KEY='djangoSecretKeytyw!b^d#elqz2d@c!m)o)g=pfnkilc07mof=pa8f@zszl5^fe=forSregistry'
...
uwsgi_1   | Wed Oct 11 15:24:11 2017 - your server socket listen backlog is limited to 100 connections
uwsgi_1   | Wed Oct 11 15:24:11 2017 - your mercy for graceful operations on workers is 60 seconds
uwsgi_1   | Wed Oct 11 15:24:11 2017 - mapped 145536 bytes (142 KB) for 1 cores
uwsgi_1   | Wed Oct 11 15:24:11 2017 - *** Operational MODE: single process ***
uwsgi_1   | Traceback (most recent call last):
uwsgi_1   |   File "shub/wsgi.py", line 7, in <module>
uwsgi_1   |     application = get_wsgi_application()
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/core/wsgi.py", line 13, in get_wsgi_application
uwsgi_1   |     django.setup(set_prefix=False)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/__init__.py", line 22, in setup
uwsgi_1   |     configure_logging(settings.LOGGING_CONFIG, settings.LOGGING)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/conf/__init__.py", line 56, in __getattr__
uwsgi_1   |     self._setup(name)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/conf/__init__.py", line 41, in _setup
uwsgi_1   |     self._wrapped = Settings(settings_module)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/conf/__init__.py", line 110, in __init__
uwsgi_1   |     mod = importlib.import_module(self.SETTINGS_MODULE)
uwsgi_1   |   File "/usr/local/lib/python3.5/importlib/__init__.py", line 126, in import_module
uwsgi_1   |     return _bootstrap._gcd_import(name[level:], package, level)
uwsgi_1   |   File "<frozen importlib._bootstrap>", line 986, in _gcd_import
uwsgi_1   |   File "<frozen importlib._bootstrap>", line 969, in _find_and_load
uwsgi_1   |   File "<frozen importlib._bootstrap>", line 944, in _find_and_load_unlocked
uwsgi_1   |   File "<frozen importlib._bootstrap>", line 222, in _call_with_frames_removed
uwsgi_1   |   File "<frozen importlib._bootstrap>", line 986, in _gcd_import
uwsgi_1   |   File "<frozen importlib._bootstrap>", line 969, in _find_and_load
uwsgi_1   |   File "<frozen importlib._bootstrap>", line 958, in _find_and_load_unlocked
uwsgi_1   |   File "<frozen importlib._bootstrap>", line 673, in _load_unlocked
uwsgi_1   |   File "<frozen importlib._bootstrap_external>", line 662, in exec_module
uwsgi_1   |   File "<frozen importlib._bootstrap>", line 222, in _call_with_frames_removed
uwsgi_1   |   File "./shub/__init__.py", line 2, in <module>
uwsgi_1   |     from .celery import shubcelery as celery_app
uwsgi_1   |   File "./shub/celery.py", line 45, in <module>
uwsgi_1   |     from opbeat.contrib.django.models import (
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/opbeat/contrib/django/models.py", line 226, in <module>
uwsgi_1   |     if 'opbeat.contrib.django' in django_settings.INSTALLED_APPS:
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/conf/__init__.py", line 56, in __getattr__
uwsgi_1   |     self._setup(name)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/conf/__init__.py", line 41, in _setup
uwsgi_1   |     self._wrapped = Settings(settings_module)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/conf/__init__.py", line 129, in __init__
uwsgi_1   |     raise ImproperlyConfigured("The SECRET_KEY setting must not be empty.")
uwsgi_1   | django.core.exceptions.ImproperlyConfigured: The SECRET_KEY setting must not be empty.
uwsgi_1   | Wed Oct 11 15:24:13 2017 - unable to load app 0 (mountpoint='') (callable not found or import error)
uwsgi_1   | Wed Oct 11 15:24:13 2017 - *** no app loaded. going in full dynamic mode ***
uwsgi_1   | Wed Oct 11 15:24:13 2017 - *** uWSGI is running in multiple interpreter mode ***
uwsgi_1   | Wed Oct 11 15:24:13 2017 - spawned uWSGI master process (pid: 47)
uwsgi_1   | Wed Oct 11 15:24:13 2017 - spawned uWSGI worker 1 (pid: 51, cores: 1)
uwsgi_1   | Wed Oct 11 15:24:13 2017 - --- no python application found, check your startup logs for errors ---
uwsgi_1   | [pid: 51|app: -1|req: -1/1] 185.62.205.145 () {38 vars in 582 bytes} [Wed Oct 11 15:24:13 2017] GET / => generated 21 bytes in 0 msecs (HTTP/1.1 500) 2 headers in 83 bytes (0 switches on core 0)
```
Big sigh.

```
moeller@server2:~/sregistry$ docker-compose stop
Stopping sregistry_nginx_1 ... done
Stopping sregistry_uwsgi_1 ... done
Stopping sregistry_redis_1 ... done
Stopping sregistry_db_1 ... done
```
Hm.

```
$ sudo useradd vsoch
$ sudo mkdir /home/vsoch
$ chown vsoch.vsoch !$
$ sudo chown vsoch.vsoch /home/vsoch
$ sudo vim /etc/group
$ grep docker /etc/group
docker:x:999:moeller,vsoch
```

Dear @vsoch, If you do not mind too much, please send me public ssh key of yours and then have a look yourself.
