### 2017-10-11 #426 @vsoch
hey @smoe ! You named your file `secret.py`. It needs to be named `secrets.py`, per the instruction!

https://singularityhub.github.io/sregistry/install.html#secrets

The reason it isn't finding it is because it looks specifically for that file, and defaults to importing "bogus_secrets.py"

We will get this running, don't worry! You are doing great! :)
