### 2017-10-11 #426 @smoe
Wow! A classic one. Many thanks for spotting that. No sure if I deserve that praise, though. Now it is the "502 Bad Gateway" and
```
uwsgi_1   | No changes detected
uwsgi_1   | Operations to perform:
uwsgi_1   |   Apply all migrations: auth
uwsgi_1   | Running migrations:
uwsgi_1   |   Applying contenttypes.0001_initial... OK
uwsgi_1   |   Applying contenttypes.0002_remove_content_type_name... OK
uwsgi_1   |   Applying auth.0001_initial... OK
uwsgi_1   |   Applying auth.0002_alter_permission_name_max_length... OK
uwsgi_1   |   Applying auth.0003_alter_user_email_max_length... OK
uwsgi_1   |   Applying auth.0004_alter_user_username_opts... OK
uwsgi_1   |   Applying auth.0005_alter_user_last_login_null... OK
uwsgi_1   |   Applying auth.0006_require_contenttypes_0002... OK
uwsgi_1   |   Applying auth.0007_alter_validators_add_error_messages... OK
uwsgi_1   |   Applying auth.0008_alter_user_username_max_length... OK
uwsgi_1   | Traceback (most recent call last):
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/backends/utils.py", line 65, in execute
uwsgi_1   |     return self.cursor.execute(sql, params)
uwsgi_1   | psycopg2.ProgrammingError: relation "users_user" does not exist
uwsgi_1   | LINE 1: ...gree_terms", "users_user"."agree_terms_date" FROM "users_use...
uwsgi_1   |                                                              ^
uwsgi_1   |
uwsgi_1   |
uwsgi_1   | The above exception was the direct cause of the following exception:
uwsgi_1   |
uwsgi_1   | Traceback (most recent call last):
uwsgi_1   |   File "manage.py", line 8, in <module>
uwsgi_1   |     execute_from_command_line(sys.argv)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/core/management/__init__.py", line 363, in execute_from_command_line
uwsgi_1   |     utility.execute()
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/core/management/__init__.py", line 355, in execute
uwsgi_1   |     self.fetch_command(subcommand).run_from_argv(self.argv)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/core/management/base.py", line 283, in run_from_argv
uwsgi_1   |     self.execute(*args, **cmd_options)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/core/management/base.py", line 330, in execute
uwsgi_1   |     output = self.handle(*args, **options)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/core/management/commands/migrate.py", line 227, in handle
uwsgi_1   |     self.verbosity, self.interactive, connection.alias, apps=post_migrate_apps, plan=plan,
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/core/management/sql.py", line 53, in emit_post_migrate_signal
uwsgi_1   |     **kwargs
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/dispatch/dispatcher.py", line 193, in send
uwsgi_1   |     for receiver in self._live_receivers(sender)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/dispatch/dispatcher.py", line 193, in <listcomp>
uwsgi_1   |     for receiver in self._live_receivers(sender)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/guardian/management/__init__.py", line 33, in create_anonymous_user
uwsgi_1   |     User.objects.get(**lookup)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/models/manager.py", line 85, in manager_method
uwsgi_1   |     return getattr(self.get_queryset(), name)(*args, **kwargs)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/models/query.py", line 374, in get
uwsgi_1   |     num = len(clone)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/models/query.py", line 232, in __len__
uwsgi_1   |     self._fetch_all()
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/models/query.py", line 1105, in _fetch_all
uwsgi_1   |     self._result_cache = list(self._iterable_class(self))
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/models/query.py", line 53, in __iter__
uwsgi_1   |     results = compiler.execute_sql(chunked_fetch=self.chunked_fetch)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/models/sql/compiler.py", line 886, in execute_sql
uwsgi_1   |     raise original_exception
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/models/sql/compiler.py", line 876, in execute_sql
uwsgi_1   |     cursor.execute(sql, params)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/backends/utils.py", line 65, in execute
uwsgi_1   |     return self.cursor.execute(sql, params)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/utils.py", line 94, in __exit__
uwsgi_1   |     six.reraise(dj_exc_type, dj_exc_value, traceback)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/utils/six.py", line 685, in reraise
uwsgi_1   |     raise value.with_traceback(tb)
uwsgi_1   |   File "/usr/local/lib/python3.5/site-packages/django/db/backends/utils.py", line 65, in execute
uwsgi_1   |     return self.cursor.execute(sql, params)
uwsgi_1   | django.db.utils.ProgrammingError: relation "users_user" does not exist
uwsgi_1   | LINE 1: ...gree_terms", "users_user"."agree_terms_date" FROM "users_use...
uwsgi_1   |                                                              ^
uwsgi_1   |
uwsgi_1   | Migrations for 'users':
uwsgi_1   |   shub/apps/users/migrations/0001_initial.py
uwsgi_1   |     - Create model User
uwsgi_1   | Migrations for 'main':
uwsgi_1   |   shub/apps/main/migrations/0001_initial.py
uwsgi_1   |     - Create model Collection
uwsgi_1   |     - Create model Container
uwsgi_1   |     - Create model Demo
uwsgi_1   |     - Create model Label
uwsgi_1   |     - Create model Share
uwsgi_1   |     - Create model Star
uwsgi_1   |     - Alter unique_together for star (1 constraint(s))
uwsgi_1   |     - Alter unique_together for share (1 constraint(s))
uwsgi_1   |     - Alter unique_together for label (1 constraint(s))
uwsgi_1   |     - Alter unique_together for container (1 constraint(s))
uwsgi_1   | Migrations for 'api':
uwsgi_1   |   shub/apps/api/migrations/0001_initial.py
uwsgi_1   |     - Create model ImageFile
uwsgi_1   | Migrations for 'logs':
uwsgi_1   |   shub/apps/logs/migrations/0001_initial.py
uwsgi_1   |     - Create model APIRequestCount
uwsgi_1   |     - Create model APIRequestLog
uwsgi_1   | Operations to perform:
uwsgi_1   |   Apply all migrations: admin, api, auth, authtoken, contenttypes, djcelery, guardian, logs, main, sessions, social_django, taggit, users
uwsgi_1   | Running migrations:
uwsgi_1   |   Applying users.0001_initial... OK
```
