### 2017-10-11 #426 @vsoch
It's so great to make your acquaintance! I think references to Singularity images is a fantastic idea. I am definitely going to check out these resources, and @gmk @godlovedc and @bauerm97 see the discussion above!
