### 2017-10-12 #426 @smoe
That thread on the Debian Med was rather fruitful  - the NeuroDebian folks created a singularity image, see https://lists.debian.org/debian-med/2017/10/msg00037.html .
