### 2017-10-15 #426 @arfon
@vsoch - please merge this PR https://github.com/singularityhub/sregistry/pull/43 and then
could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
