### 2017-10-15 #426 @vsoch
Thanks @arfon! Here is the completed DOI, which I've added additionally to the README.md:

https://doi.org/10.5281/zenodo.1012531
